Teach hangouts-notifier about the chat frames in gmail
Even if we don't find a message, keep pooling frames with 'prop=gmail#epreld' in the URL.
All I can say is it works for me. It's dirty (maybe too much to merge), seems too brittle and clings to one non-chat iframe. It may or may not be extended in the same fashion for other google apps that connect to hangouts.